requirements.txt
astroid==1.6.0
attrdict==2.0.0
attrs==17.4.0
Automat==0.6.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
basicstruct==1.0.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
constantly==15.1.0
coverage==4.5.1
decorator==4.1.2
Django==1.7.11
docopt==0.6.2
docutils==0.14
enum34==1.1.6
flake8==3.5.0
functools32==3.2.3.post2
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipdb==0.10.3
ipdbugger==1.1.2
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.2.15
jsonschema==2.6.0
lazy-object-proxy==1.3.1
lockfile==0.12.2
lxml==3.8.0
mccabe==0.6.1
mock==2.0.0
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
prompt-toolkit==1.0.15
psutil==5.4.3
ptyprocess==0.5.2
pycodestyle==2.3.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.1
pytest==3.4.1
pytest-cov==2.5.1
python-daemon==2.1.2
PyYAML==3.12
reportportal-client==3.0.0
requests==2.18.4
rotest==2.4.1
scandir==1.6
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
traitlets==4.3.2
Twisted==17.9.0
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
xlwt==1.3.0
xmltodict==0.11.0
zope.interface==4.4.3
